refactor(#18): Renamed `borrowtracker` to `borrow_tracker` to stay consistent w/ Miri
